feat: change logback config, log request info
