+++ api/build.gradle.kts
@@ -15,4 +15,6 @@ dependencies {
 
     implementation("com.google.firebase:firebase-admin:7.1.0")
     implementation("org.springdoc:springdoc-openapi-ui:1.5.+")
+
+    compileOnly("ch.qos.logback:logback-classic")
 }
